pyproject.toml
[project]
name = "signal-agent"
version = "0.1.0"
description = "Arthur AI Governance Signal Agent — Phase 1 MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.2",
    "httpx>=0.27",
    "anthropic>=0.39",
    "hubspot-api-client>=11.1",
    "slack-sdk>=3.33",
    "inngest>=0.4",
    "pyyaml>=6.0",
    "rapidfuzz>=3.10",
    "tenacity>=9.0",
    "structlog>=24.4",
    "beautifulsoup4>=4.12",
    "playwright>=1.48",
    # Arthur tracing (OpenTelemetry + OpenInference). The Arthur Engine accepts
    # OTLP/HTTP on /v1/traces with a Bearer token.
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
    "openinference-instrumentation-anthropic>=0.1.8",
    "opentelemetry-instrumentation-httpx>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["signal_agent*"]
exclude = ["migrations*", "tests*", "docs*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4", "SIM"]
ignore = ["S101"]  # allow asserts in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
